pyproject.toml
[project]
name = "f1-realtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.1.2",
    "livef1>=0.1.50",
    "loguru>=0.7.3",
    "pinotdb>=5.6.0",
    "plotly>=6.0.1",
    "streamlit>=1.44.1",
]
